refactor(dashboard): collapse toolset detail surfaces into MCP server pages
The toolset-specific listing card, table row, and 2,100-line detail page
duplicated the MCP server components; the server implementations are now
the single path, driven by backendKind and toolsetSummary. Toolset-owned
concerns (tool selection, resources, prompts, OAuth, performance mode)
appear as conditional tabs on the unified /mcp/x detail surface, while
name, visibility, publishing, instructions, and addressing write through
the mcpServers/mcpEndpoints management APIs. Old /mcp/:toolsetSlug detail
URLs redirect tab-for-tab to the wrapper's server page so existing
bookmarks keep working.
