--- client/dashboard/src/components/mcp/MCPCard.tsx
@@ -1,191 +1,0 @@
-import { CopyButton } from "@/components/ui/CopyButton";
-import { DotCard } from "@/components/ui/DotCard";
-import { Button } from "@/components/ui/Button";
-import { Text } from "@/components/ui/Text";
-import { useMcpUrl } from "@/hooks/useToolsetUrl";
-import { useRoutes } from "@/routes";
-import { MCPStatusIndicator } from "./MCPStatusIndicator";
-import { MCPActivityIndicator } from "./MCPActivityIndicator";
-import type { McpActivityStatus } from "./mcp-activity";
-import { ToolsetEntry } from "@gram/client/models/components/toolsetentry.js";
-import { useLatestDeployment } from "@gram/client/react-query/latestDeployment.js";
-import {
-  AlertTriangleIcon,
-  ArrowRight,
-  Link2,
-  Network,
-  Package,
-} from "lucide-react";
-import { useMemo } from "react";
-import {
-  useCatalogIconMap,
-  useExternalMcpOAuthConfigStatus,
-} from "../sources/sources-hooks";
-import { ToolCollectionBadge } from "../tool-collection-badge";
-import { Badge } from "@/components/ui/Badge";
-
-export function MCPCard({
-  toolset,
-  activityStatus,
-  recentWindowDays,
-}: {
-  toolset: ToolsetEntry;
-  activityStatus?: McpActivityStatus | null;
-  recentWindowDays?: number;
-}): JSX.Element {
-  const routes = useRoutes();
-  const { installPageUrl } = useMcpUrl(toolset);
-  const catalogIconMap = useCatalogIconMap();
-  const { data: deploymentResult } = useLatestDeployment();
-  const oauthStatus = useExternalMcpOAuthConfigStatus(toolset.slug);
-
-  const externalMcpLogoUrl = useMemo(() => {
-    const externalMcpUrn = toolset.toolUrns?.find((urn) =>
-      urn.includes(":externalmcp:"),
-    );
-    const slug = externalMcpUrn?.split(":")[2];
-    if (!slug) return undefined;
-
-    const matchingMcp = deploymentResult?.deployment?.externalMcps?.find(
-      (mcp) => mcp.slug === slug,
-    );
-    return matchingMcp?.registryServerSpecifier
-      ? catalogIconMap.get(matchingMcp.registryServerSpecifier)
-      : undefined;
-  }, [toolset.toolUrns, catalogIconMap, deploymentResult]);
-
-  const handleClick = () => {
-    if (oauthStatus === "required-unconfigured") {
-      routes.mcp.details.authentication.goTo(toolset.slug);
-    } else {
-      routes.mcp.details.goTo(toolset.slug);
-    }
-  };
-
-  const installSourceTooltip = toolset.origin?.registrySpecifier
-    ? `Installed from ${toolset.origin.registrySpecifier}`
-    : undefined;
-
-  // External MCP "proxy" servers can't enumerate their tools until a user
-  // authenticates against them, so hide the misleading "No Tools" badge and
-  // surface the visible (non-proxy) tools only.
-  const visibleToolNames = toolset.tools
-    .filter((t) => !(t.type === "externalmcp" && t.name.endsWith(":proxy")))
-    .map((t) => t.name);
-  const isExternalMcpProxy = toolset.tools.some(
-    (t) => t.type === "externalmcp" && t.name.endsWith(":proxy"),
-  );
-
-  return (
-    <DotCard
-      className="cursor-pointer"
-      onClick={handleClick}
-      overlay={
-        oauthStatus === "required-unconfigured" && (
-          <div className="absolute bottom-3.5 left-1/2 z-10 -translate-x-1/2">
-            <Badge variant="warning">
-              <Badge.LeftIcon>
-                <AlertTriangleIcon />
-              </Badge.LeftIcon>
-              <Badge.Text>OAuth Required</Badge.Text>
-            </Badge>
-          </div>
-        )
-      }
-      icon={
-        externalMcpLogoUrl ? (
-          <img
-            src={externalMcpLogoUrl}
-            alt={toolset.name}
-            className="h-12 w-12 object-contain"
-          />
-        ) : (
-          <Network className="text-muted-foreground h-8 w-8" />
-        )
-      }
-    >
-      {/* Header row with name */}
-      <div className="mb-2 flex items-start justify-between gap-2">
-        <Text
-          variant="subheading"
-          as="div"
-          className="text-md group-hover:text-primary flex-1 truncate transition-colors"
-          title={toolset.name}
-        >
-          {toolset.name}
-        </Text>
-        <div className="flex items-center gap-1">
-          {installPageUrl && (
-            <CopyButton
-              text={installPageUrl}
-              size="sm"
-              icon={Link2}
-              tooltip="Copy install page URL"
-            />
-          )}
-          {installSourceTooltip && (
-            <Button
-              type="button"
-              variant="tertiary"
-              size="sm"
-              tooltip={installSourceTooltip}
-              aria-label={installSourceTooltip}
-              onClick={(e) => e.stopPropagation()}
-            >
-              <Package className="text-muted-foreground group-hover:text-foreground h-4 w-4" />
-            </Button>
-          )}
-          <ToolCollectionBadge
-            toolNames={visibleToolNames}
-            emptyLabel={isExternalMcpProxy ? null : undefined}
-          />
-        </div>
-      </div>
-
-      {/* Footer row with status indicator and open link */}
-      <div className="mt-auto flex items-center justify-between gap-2 pt-2">
-        <div className="flex items-center gap-2">
-          <MCPStatusIndicator
-            mcpEnabled={toolset.mcpEnabled}
-            mcpIsPublic={toolset.mcpIsPublic}
-          />
-          {activityStatus && (
-            <MCPActivityIndicator
-              status={activityStatus}
-              recentWindowDays={recentWindowDays}
-            />
-          )}
-        </div>
-        {oauthStatus === "required-unconfigured" ? (
-          <div className="text-warning flex items-center gap-1 text-sm">
-            <span>Set up</span>
-            <ArrowRight className="h-3.5 w-3.5" />
-          </div>
-        ) : (
-          <div className="text-muted-foreground group-hover:text-primary flex items-center gap-1 text-sm transition-colors">
-            <span>Open</span>
-            <ArrowRight className="h-3.5 w-3.5" />
-          </div>
-        )}
-      </div>
-    </DotCard>
-  );
-}
-
-export function MCPCardSkeleton(): JSX.Element {
-  return (
-    <DotCard>
-      <div className="mb-2 flex items-start justify-between gap-2">
-        <div className="bg-muted h-5 w-2/3 animate-pulse rounded" />
-        <div className="bg-muted h-5 w-10 animate-pulse rounded-full" />
-      </div>
-      <div className="mt-auto flex items-center justify-between gap-2 pt-2">
-        <div className="flex items-center gap-2">
-          <div className="bg-muted h-2.5 w-2.5 animate-pulse rounded-full" />
-          <div className="bg-muted h-3.5 w-12 animate-pulse rounded" />
-        </div>
-        <div className="bg-muted h-3.5 w-10 animate-pulse rounded" />
-      </div>
-    </DotCard>
-  );
-}
