+++ client/dashboard/src/pages/mcp/x/tabs/settings/sections/ToolFilteringSection.tsx
@@ -7,15 +7,18 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/Select";
+import { Toolset } from "@/lib/toolTypes";
 import { toolVariationsGroupDisplayName } from "@/lib/toolVariationGroups";
 import type { McpServer } from "@gram/client/models/components/mcpserver.js";
 import { useCreateGlobalToolVariationGroupMutation } from "@gram/client/react-query/createGlobalToolVariationGroup.js";
 import { invalidateAllGetMcpServer } from "@gram/client/react-query/getMcpServer.js";
 import { invalidateAllMcpServers } from "@gram/client/react-query/mcpServers.js";
+import { useSetToolsetToolVariationsGroupMutation } from "@gram/client/react-query/setToolsetToolVariationsGroup.js";
 import {
   invalidateAllToolVariationGroups,
   useToolVariationGroups,
 } from "@gram/client/react-query/toolVariationGroups.js";
+import { invalidateAllToolset } from "@gram/client/react-query/toolset.js";
 import { useUpdateMcpServerMutation } from "@gram/client/react-query/updateMcpServer.js";
 import { Button } from "@/components/ui/Button";
 import { useQueryClient } from "@tanstack/react-query";
@@ -30,14 +33,27 @@ const DISABLED_VALUE = "__disabled__";
 
 export function ToolFilteringSection({
   mcpServer,
+  backingToolset,
 }: {
   mcpServer: McpServer;
+  /**
+   * When set, the variations group lives on the backing toolset: reads come
+   * from — and writes go to — the toolset, matching where existing hosted
+   * servers keep their filtering config. The server-side runtime resolves the
+   * wrapper's own group first and falls back to the toolset's.
+   */
+  backingToolset?: Toolset;
 }): JSX.Element {
   const queryClient = useQueryClient();
   const groupsQuery = useToolVariationGroups();
   const groups = groupsQuery.data?.groups ?? [];
 
-  const currentValue = mcpServer.toolVariationsGroupId ?? DISABLED_VALUE;
+  const persistedGroupId = backingToolset
+    ? (mcpServer.toolVariationsGroupId ??
+      backingToolset.toolVariationsGroupId ??
+      null)
+    : (mcpServer.toolVariationsGroupId ?? null);
+  const currentValue = persistedGroupId ?? DISABLED_VALUE;
   const [draft, setDraft] = useState(currentValue);
 
   // Re-sync the draft when the persisted value changes underneath us.
@@ -63,9 +79,33 @@ export function ToolFilteringSection({
     onError: notifyError,
   });
 
-  // mcpServers.update is a full-record replace for the optional UUID
-  // references, so every field has to be re-sent or it gets nulled.
+  const setToolsetGroup = useSetToolsetToolVariationsGroupMutation({
+    onSuccess: async () => {
+      await Promise.all([
+        invalidateAllToolset(queryClient, { refetchType: "all" }),
+        invalidateAllGetMcpServer(queryClient, { refetchType: "all" }),
+        invalidateAllMcpServers(queryClient, { refetchType: "all" }),
+      ]);
+      toast.success("Tool filtering settings updated");
+    },
+    onError: notifyError,
+  });
+
   const applyGroup = (groupId: string | null) => {
+    if (backingToolset) {
+      setToolsetGroup.mutate({
+        request: {
+          slug: backingToolset.slug,
+          setToolVariationsGroupRequestBody: {
+            toolVariationsGroupId: groupId ?? undefined,
+          },
+        },
+      });
+      return;
+    }
+
+    // mcpServers.update is a full-record replace for the optional UUID
+    // references, so every field has to be re-sent or it gets nulled.
     updateMcpServer.mutate({
       request: {
         updateMcpServerForm: {
@@ -98,7 +138,10 @@ export function ToolFilteringSection({
       ),
   });
 
-  const isSaving = updateMcpServer.isPending || createGroup.isPending;
+  const isSaving =
+    updateMcpServer.isPending ||
+    setToolsetGroup.isPending ||
+    createGroup.isPending;
   const dirty = draft !== currentValue;
   const hasGroups = groups.length > 0;
 
@@ -170,7 +213,9 @@ export function ToolFilteringSection({
             <SettingsSection.FooterActions>
               <RequireScope scope="mcp:write" level="component">
                 <FooterSaveButton
-                  pending={updateMcpServer.isPending}
+                  pending={
+                    updateMcpServer.isPending || setToolsetGroup.isPending
+                  }
                   disabled={!dirty || isSaving}
                   onClick={() =>
                     applyGroup(draft === DISABLED_VALUE ? null : draft)
