+++ client/dashboard/src/pages/sources/tunneled-mcp/TunneledMCPDetails.tsx
@@ -25,6 +25,7 @@ import {
 } from "@/lib/sources";
 import { TUNNELED_MCP_FEATURE_FLAG } from "@/lib/tunneledMcp";
 import { useRoutes } from "@/routes";
+import type { McpEndpoint } from "@gram/client/models/components/mcpendpoint.js";
 import type { McpServer } from "@gram/client/models/components/mcpserver.js";
 import type { TunneledMcpConnection } from "@gram/client/models/components/tunneledmcpconnection.js";
 import type { TunneledMcpServer } from "@gram/client/models/components/tunneledmcpserver.js";
@@ -137,18 +138,18 @@ function TunneledMCPDetailsContent(): JSX.Element {
     tunneledMcpServerId,
   );
 
-  const { data: endpointsResult } = useMcpEndpoints({}, undefined, {
-    enabled: tunneledMcpServerId !== "",
-  });
-  const endpointCountByServerId = useMemo(() => {
-    const counts = new Map<string, number>();
+  const { data: endpointsResult, isLoading: isLoadingEndpoints } =
+    useMcpEndpoints({}, undefined, {
+      enabled: tunneledMcpServerId !== "",
+    });
+  const endpointsByServerId = useMemo(() => {
+    const grouped = new Map<string, McpEndpoint[]>();
     for (const endpoint of endpointsResult?.mcpEndpoints ?? []) {
-      counts.set(
-        endpoint.mcpServerId,
-        (counts.get(endpoint.mcpServerId) ?? 0) + 1,
-      );
+      const existing = grouped.get(endpoint.mcpServerId);
+      if (existing) existing.push(endpoint);
+      else grouped.set(endpoint.mcpServerId, [endpoint]);
     }
-    return counts;
+    return grouped;
   }, [endpointsResult]);
 
   if (isError || (!isLoading && !tunneledMcpServer)) {
@@ -230,7 +231,8 @@ function TunneledMCPDetailsContent(): JSX.Element {
             <McpServersTab
               isLoading={isLoadingMcpServers}
               mcpServers={linkedMcpServers}
-              endpointCountByServerId={endpointCountByServerId}
+              endpointsByServerId={endpointsByServerId}
+              isLoadingEndpoints={isLoadingEndpoints}
               tunneledMcpServer={tunneledMcpServer}
             />
           </TabsContent>
@@ -548,12 +550,14 @@ function InfoPair({
 function McpServersTab({
   isLoading,
   mcpServers,
-  endpointCountByServerId,
+  endpointsByServerId,
+  isLoadingEndpoints,
   tunneledMcpServer,
 }: {
   isLoading: boolean;
   mcpServers: McpServer[];
-  endpointCountByServerId: Map<string, number>;
+  endpointsByServerId: Map<string, McpEndpoint[]>;
+  isLoadingEndpoints: boolean;
   tunneledMcpServer: TunneledMcpServer | undefined;
 }) {
   return (
@@ -566,7 +570,8 @@ function McpServersTab({
             <MCPServerCard
               key={server.id}
               server={server}
-              endpointCount={endpointCountByServerId.get(server.id) ?? 0}
+              endpoints={endpointsByServerId.get(server.id) ?? []}
+              isLoadingEndpoints={isLoadingEndpoints}
             />
           ))}
         </div>
